% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{parse_css_file}
\alias{parse_css_file}
\title{Parse a CSS file into a data frame}
\usage{
parse_css_file(fil)
}
\arguments{
\item{fil}{path to CSS file (will be \code{\link[=path.expand]{path.expand()}}ed)}
}
\description{
Parse a CSS file into a data frame
}
