% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{parse_css_text}
\alias{parse_css_text}
\title{Parse CSS text intoa data frame}
\usage{
parse_css_text(css)
}
\arguments{
\item{css}{a character vector of CSS fulrs}
}
\description{
Parse CSS text intoa data frame
}
