% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsuka-package.R
\docType{package}
\name{tsuka}
\alias{tsuka}
\alias{tsuka-package}
\title{Parse Cascading Style Sheets}
\description{
Cascading style sheets ("CSS") is a style sheet language used for
describing the presentation of a document written in a markup language like
HTML, and is designed to enable the separation of presentation and content,
including layout, colors, and fonts. Tools are provided to parse CSS content
into data frames.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://gitlab.com/hrbrmstr/tsuka}
  \item Report bugs at \url{https://gitlab.com/hrbrmstr/tsuka/issues}
}

}
\author{
Bob Rudis (bob@rud.is)
}
\keyword{internal}
