% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsuka-package.R
\docType{package}
\name{tsuka}
\alias{tsuka}
\alias{tsuka-package}
\title{...}
\description{
\itemize{
\item URL: \url{https://gitlab.com/hrbrmstr/tsuka}
\item BugReports: \url{https://gitlab.com/hrbrmstr/tsuka/issues}
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://gitlab.com/hrbrmstr/tsuka}
  \item Report bugs at \url{https://gitlab.com/hrbrmstr/tsuka/issues}
}

}
\author{
Bob Rudis (bob@rud.is)
}
\keyword{internal}
